requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==42.0.2
distlib==0.3.8
Django==4.2.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
filelock==3.13.1
mysqlclient==2.2.4
pip==24.0
platformdirs==4.2.0
pycparser==2.21
pytz==2024.1
setuptools==57.4.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
virtualenv==20.25.0
